+++ packages/connectivity/android/src/main/java/io/flutter/plugins/connectivity/ConnectivityPlugin.java
@@ -25,165 +25,17 @@ import io.flutter.plugin.common.MethodChannel.Result;
 import io.flutter.plugin.common.PluginRegistry.Registrar;
 
 /** ConnectivityPlugin */
-public class ConnectivityPlugin implements MethodCallHandler, StreamHandler {
-  private final Registrar registrar;
-  private final ConnectivityManager manager;
-  private BroadcastReceiver receiver;
+public class ConnectivityPlugin {
 
   /** Plugin registration. */
   public static void registerWith(Registrar registrar) {
     final MethodChannel channel =
-        new MethodChannel(registrar.messenger(), "plugins.flutter.io/connectivity");
+            new MethodChannel(registrar.messenger(), "plugins.flutter.io/connectivity");
     final EventChannel eventChannel =
-        new EventChannel(registrar.messenger(), "plugins.flutter.io/connectivity_status");
-    ConnectivityPlugin instance = new ConnectivityPlugin(registrar);
-    channel.setMethodCallHandler(instance);
-    eventChannel.setStreamHandler(instance);
-  }
+            new EventChannel(registrar.messenger(), "plugins.flutter.io/connectivity_status");
 
-  private ConnectivityPlugin(Registrar registrar) {
-    this.registrar = registrar;
-    this.manager =
-        (ConnectivityManager)
-            registrar
-                .context()
-                .getApplicationContext()
-                .getSystemService(Context.CONNECTIVITY_SERVICE);
-  }
-
-  @Override
-  public void onListen(Object arguments, EventSink events) {
-    receiver = createReceiver(events);
-    registrar
-        .context()
-        .registerReceiver(receiver, new IntentFilter(ConnectivityManager.CONNECTIVITY_ACTION));
-  }
-
-  @Override
-  public void onCancel(Object arguments) {
-    registrar.context().unregisterReceiver(receiver);
-    receiver = null;
-  }
-
-  private String getNetworkType(ConnectivityManager manager) {
-    if (android.os.Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-      Network network = manager.getActiveNetwork();
-      NetworkCapabilities capabilities = manager.getNetworkCapabilities(network);
-      if (capabilities == null) {
-        return "none";
-      }
-      if (capabilities.hasTransport(NetworkCapabilities.TRANSPORT_WIFI)
-          || capabilities.hasTransport(NetworkCapabilities.TRANSPORT_ETHERNET)) {
-        return "wifi";
-      }
-      if (capabilities.hasTransport(NetworkCapabilities.TRANSPORT_CELLULAR)) {
-        return "mobile";
-      }
-    }
-
-    return getNetworkTypeLegacy(manager);
-  }
-
-  @SuppressWarnings("deprecation")
-  private String getNetworkTypeLegacy(ConnectivityManager manager) {
-    // handle type for Android versions less than Android 9
-    NetworkInfo info = manager.getActiveNetworkInfo();
-    if (info == null || !info.isConnected()) {
-      return "none";
-    }
-    int type = info.getType();
-    switch (type) {
-      case ConnectivityManager.TYPE_ETHERNET:
-      case ConnectivityManager.TYPE_WIFI:
-      case ConnectivityManager.TYPE_WIMAX:
-        return "wifi";
-      case ConnectivityManager.TYPE_MOBILE:
-      case ConnectivityManager.TYPE_MOBILE_DUN:
-      case ConnectivityManager.TYPE_MOBILE_HIPRI:
-        return "mobile";
-      default:
-        return "none";
-    }
-  }
-
-  @Override
-  public void onMethodCall(MethodCall call, Result result) {
-    switch (call.method) {
-      case "check":
-        handleCheck(call, result);
-        break;
-      case "wifiName":
-        handleWifiName(call, result);
-        break;
-      case "wifiBSSID":
-        handleBSSID(call, result);
-        break;
-      case "wifiIPAddress":
-        handleWifiIPAddress(call, result);
-        break;
-      default:
-        result.notImplemented();
-        break;
-    }
-  }
-
-  private void handleCheck(MethodCall call, final Result result) {
-    result.success(checkNetworkType());
-  }
-
-  private String checkNetworkType() {
-    return getNetworkType(manager);
-  }
-
-  private WifiInfo getWifiInfo() {
-    WifiManager wifiManager =
-        (WifiManager)
-            registrar.context().getApplicationContext().getSystemService(Context.WIFI_SERVICE);
-    return wifiManager == null ? null : wifiManager.getConnectionInfo();
-  }
-
-  private void handleWifiName(MethodCall call, final Result result) {
-    WifiInfo wifiInfo = getWifiInfo();
-    String ssid = null;
-    if (wifiInfo != null) ssid = wifiInfo.getSSID();
-    if (ssid != null) ssid = ssid.replaceAll("\"", ""); // Android returns "SSID"
-    result.success(ssid);
-  }
-
-  private void handleBSSID(MethodCall call, MethodChannel.Result result) {
-    WifiInfo wifiInfo = getWifiInfo();
-    String bssid = null;
-    if (wifiInfo != null) bssid = wifiInfo.getBSSID();
-    result.success(bssid);
-  }
-
-  private void handleWifiIPAddress(MethodCall call, final Result result) {
-    WifiManager wifiManager =
-        (WifiManager)
-            registrar.context().getApplicationContext().getSystemService(Context.WIFI_SERVICE);
-
-    WifiInfo wifiInfo = null;
-    if (wifiManager != null) wifiInfo = wifiManager.getConnectionInfo();
-
-    String ip = null;
-    int i_ip = 0;
-    if (wifiInfo != null) i_ip = wifiInfo.getIpAddress();
-
-    if (i_ip != 0)
-      ip =
-          String.format(
-              "%d.%d.%d.%d",
-              (i_ip & 0xff), (i_ip >> 8 & 0xff), (i_ip >> 16 & 0xff), (i_ip >> 24 & 0xff));
-
-    result.success(ip);
-  }
-
-  private BroadcastReceiver createReceiver(final EventSink events) {
-    return new BroadcastReceiver() {
-      @Override
-      public void onReceive(Context context, Intent intent) {
-        events.success(checkNetworkType());
-      }
-    };
+    ConnectivityMethodChannelHandler methodChannelHandler = new ConnectivityMethodChannelHandler(registrar.context());
+    channel.setMethodCallHandler(methodChannelHandler);
+    eventChannel.setStreamHandler(methodChannelHandler);
   }
 }
