+++ packages/connectivity/example/android/app/src/main/java/dev/flutter/plugins/connectivity_exmaple/MainActivity.java
@@ -1,0 +1,12 @@
+package dev.flutter.plugins.connectivity_exmaple;
+
+import io.flutter.embedding.android.FlutterActivity;
+import io.flutter.embedding.engine.FlutterEngine;
+
+public class MainActivity extends FlutterActivity {
+
+    @Override
+    public void configureFlutterEngine(FlutterEngine flutterEngine) {
+        super.configureFlutterEngine(flutterEngine);
+    }
+}
